Move implementation logic from LLHDOps.h to cpp file
